[FAB-5155]Fix README.md in "balance-transfer" example
The README in "balance-transfer" should be fixed in the section:
https://github.com/hyperledger/fabric-samples/tree/master/
balance-transfer#running-the-sample-program

1. Broken link
There is a link to "Sample REST APIs Requests" which is broken.
This is linked to non-existent path on fabric-sdk-node repository.
I believe the intended link is to
https://github.com/hyperledger/fabric-samples/tree/master/
balance-transfer#sample-rest-apis-requests

2. Instruction to start node application.
There is no instruction to start node application.
If we execute the REST APIs according to the current README,
we can't get any responses since node application is not working.
We need to execute the following steps before calling REST APIs.

 - npm install
 - PORT=4000 node app

Change-Id: I9ec0ec18bb36addd6c8371481dc8cfe645053ec6
Signed-off-by: Yuki Kondo <yuki.kondo@hal.hitachi.com>