[FAB-5024]Failed to find a peer matching localhost:8051
Connecting to event hubs is not implemented correctly
in helper.js of balance-transfer.

When an application looks for event hubs, this app can only
connect to event hubs in its own org. This check is executed in
loop processing. If the org name of requester and peer is
different, we should go to next loop (i.e. next peer).

Current code doesn't go back to the head of the loop correctly.
As a result, incorrect error message is submitted.

This patch fixes the break point to go back to the head of the
loop correctly.

Change-Id: Id8c9220e80c04570e90cbb74add95546a312a7af
Signed-off-by: Yuki Kondo <yuki.kondo@hal.hitachi.com>