[FAB-5285]Fix instantiate chaincode in balance-transfer
In balance-transfer, the document says "instantiate Chaincode"
will instantiate the chaincode on peer0 of org1,
but actually peer1 of org1 is also instantiated.

The reason is that instantiate-chaincode.js doesn’t have
mechanism to select target peer. Transaction proposal is sent to
all peers which is set by helper.setupPeers.

This patch fixes the code to accept target peer from REST API.

Change-Id: I17781e2d1aaf50a8c6c2db5704fb2ef6329a65a8
Signed-off-by: Yuki Kondo <yuki.kondo@hal.hitachi.com>