[FAB-6606]Fix caller name in logs of balance-transfer
In balance-transfer, a function call “getOrgAdmin(org)”
outputs incorrect user name in error message.

The following functions are executed by Admin due to
access control.
- createChannel
- JoinChannel
- installChaincode
- instantiateChaincode
- getInstalledChaincodes

In these functions, helper.getOrgAdmin(org) is called to
get enrollment for Admins. However, non-admin user name is
printed out in error messages if getOrgAdmin(org) fails. Error
messages should be modified to show that enrolling Admin fails.

In addition, an arg "username" is not used in functions above
because caller name is assigned base on org name.
To make it easy to understand the code structure,
the arg "username" should be removed.

This patch modifies caller name to Admin.
In addition, unused "username" was removed.

Change-Id: I9cf1129ae68657250de1f4b885bf29bf2c28c66e
Signed-off-by: Yuki Kondo <yuki.kondo@hal.hitachi.com>