[FAB-8540] Add ledger persistance to first-network
This CR is for the release branch.

This CR adds a new docker-compose-persist.yaml file which creates a
volume mount in the current directory at ./ledgers/<container> and binds
it to the ledger directory for the container.

It also adds a '-p' flag to the byfn.sh script to allow users to include
this new compose file when bringing up the network.

Change-Id: I311ca1d2157dd87e61ad890c98cde7c118d5ee7d
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>