[FAB-8540] Add ability byfn.sh to persist ledgers
This CR adds a new docker-compose-persist.yaml file which creates a
volume mount in the current directory at ./ledgers/<container> and binds
it to the ledger directory for the container.

It also adds a '-p' flag to the byfn.sh script to allow users to include
this new compose file when bringing up the network.

Change-Id: I89e9db111b908c9aea07715f992c79b45d13d939
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>