[FAB-8567] Remove explicit persistence path (m)
Master branch

This CR changes the ledger persistence from being to the local directory
to being managed by docker-compose.  Having docker-compose manage the
volumes is probably for the best overall, though it does make detecting
the persisted ledgers a little more challenging.

This change is required because the '-p' fails under Vagrant when
executed from within the volume which is shared with the host.

Change-Id: I9c84e8e4e557b3522ba9f3b0c6b8460b9f02f39c
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>