[FAB-8567] Remove explicit persistence path (r)
For release branch

This CR changes the ledger persistence from being to the local directory
to being managed by docker-compose.  Having docker-compose manage the
volumes is probably for the best overall, though it does make detecting
the persisted ledgers a little more challenging.

This change is required because the '-p' fails under Vagrant when
executed from within the volume which is shared with the host.

Change-Id: Ic1ac959e83043d8105c80c24b0e89b346c148b3c
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>