[FAB-8567] Alt: Always use volumes for ledger
Docker compose makes volume management easy and transparent to the user.
By defaulting the ledgers to be docker compose volumes, we transparently
get ledger persistence for the ugprade case, and may trivially tear the
volumes down in the docker-compose down command with the addition of the
--volumes flag.  This would allow the upgrade scenario to work withou
the use of the '-p' flag, and is likely to be generally much cleaner.

Change-Id: Id008d50487e8d304d6e2517300402dfd1cc8c0c3
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>