[FAB-9033] TypeScript sample uses latest type defs
This patchset makes the TypeScript version of balance-transfer
use the type definitions in the npm repository, and removes
its own type definitions.

As reported in FAB-8942, the type information in v1.1 is broken,
so temporarily it depends on an unstable (snapshot) version of
fabric-client and fabric-ca-client.
It should be fixed later once the next minor update is released.

Change-Id: I06d5886a2c7a6c932b286690478708b994e6cfb3
Signed-off-by: Taku Shimosawa <taku.shimosawa@hal.hitachi.com>