[FAB-8958] Add org3 removal to byfn.sh
When running the sample "Adding an Org to a Channel", then upon clean up,
./eyfn.sh -m down script must be run before ./byfn.sh -m down; otherwise,
org3 containers and volumes and containers do not get cleaned properly. This
patch adds the org3 container and volume clean up to ./byfn.sh so that a
user does not have to worry about the correct ordering of ./eyfn.sh and
./byfn.sh scripts.

Commented the new docker-compose due to jenkin failures. testing...

Change-Id: Ieac1adc4081d59943286cfe9139a74395986460b
Signed-off-by: Salman Baset <sabaset@us.ibm.com>