FAB-11330 Replaced 'unstabled' version with '1.2.x'
The 1.2 release should have referenced 1.2 version of the SDK.
But it used "unstable". As a result, testAPI.sh fails because
client.getEventHubForOrg() has been removed in master.

This fix is meant for patching the 1.2 branch.
I think for this we first need to create a hotfix branch. But want
to discuss with you guys first how you'd like to handle it.

Change-Id: Id90186fd1310184f4da8318d540005ebcd14221d
Signed-off-by: Jim Zhang <jim_the_matrix@hotmail.com>