FAB-11493 Rework High-Throughput tutorial
There are a couple of issues with the high-throughput tutorial that
make it difficult to follow.

This CR would make the README clearer,
as well as provided a small update to how it works.

Changes include creating a 'delstandard' invoke call in high-throughput.go
which allows standard assets that have been put on-chain to be
removed with an invoke call, which is also created as part of
the 'del-traditional.sh' script.

Another small change is that the chaincode itself has been moved to
it's own folder within fabric-samples/chaincode.  It does not make sense
to have this chaincode sitting on it's own; it should be located where
the other chaincodes are to ease confusion.

In order to test these changes, I ran the entire tutorial from scratch
based on what I detailed in the README, which worked perfectly.

Change-Id: Ie90e362040df97b0eae46c1802c925ac7587d882
Signed-off-by: Ethan Fox <Ethan.Fox@ibm.com>