[FAB-12371]Fix the abac sample to use new cid package
The latest cid package is fabric/core/chaincode/shim/ext/cid.
This CR fixes the abac sample to use the new cid package instead
of the old one at fabric/core/chaincode/lib/cid.
The vendor folder is removed because the cid package was moved
down to the shim package. Chaincode doesn't need to prepare
dependencies for the cid and related packages.

Change-Id: Ifc431ee4504cd009f84f9d5de271721597bba6e7
Signed-off-by: Yuki Kondo <yuki.kondo@hal.hitachi.com>