[FAB-12880] Remove old prog model samples for FabCar
Remove the old programming model samples for FabCar,
including the "node" (JavaScript!) chaincode sample
and the JavaScript application in the top level.

At the same time, update the startFabric.sh script
to output correct help information on how to run
the applications after starting the Fabric network
and deploying the chaincode.

Change-Id: I1e3701737f105cba67f661bf893771ac8a3a0373
Signed-off-by: Simon Stone <sstone1@uk.ibm.com>