[FAB-16210] Add script to install chaincode peer1.org1
The first-network consists of four nodes. However, since
the byfn script installs chaincode for test only on 3 peers.

So, an error occurs when invoking the 4 peers from the HLF client.
Of course, you do not need to get success from 4 peers,
but I think you should install all for the convenience of the beginners.

Change-Id: Ie236a44d607eccc8e5bac0b7d9686b44810cabec

Signed-off-by: amuyu <hwajoong.song@gmail.com>
Change-Id: I3ed481f1e7939b3f80530860f944a5549492cb59