[FAB-16210] Add script to install chaincode peer1.org1
The first-network consists of four nodes. However, since
the byfn script installs chaincode for test only on 3 peers.

So, an error occurs when invoking the 4 peers from the HLF client.
Of course, you do not need to get success from 4 peers,
but I think you should install all for the convenience of the beginners.

Signed-off-by: amuyu <hwajoong.song@gmail.com>
Change-Id: I21f54814d09c33c66a4e80c2ab96a4b067c1e753